task 5 aki mungu saidia

+++ 0x0C-more_malloc_free/101-mul.c
@@ -4,7 +4,7 @@
 
 /**
   * main - multiplies two positive numbers
-  * @argc: number of arguments
+  * @argc: n arguments
   * @argv: args
   * Return: int
   */
@@ -25,7 +25,7 @@ int i, j;
 		}
 
 	}
-	mul = atoi(argv[1]) *  atoi(argv[2]);
+	mul = atol(argv[1]) *  atol(argv[2]);
 	printf("%lu\n", mul);
 return (0);
 }
